Loader carga un OpenHash de MovieRating en un Openhash, hay que modificarlo para que sea un heap. Lo hare despues del almuerzo

+++ src/Modelo/Rating.java
@@ -2,7 +2,7 @@ package Modelo;
 
 public class Rating {
 
-    private float NumberVotes;
-    private float AverageRating;
+    private Integer NumberVotes;
+    private Float AverageRating;
 
 }
